+++ api/model/model.go
@@ -1,10 +1,10 @@
 package model
 
 type DemoOrder struct {
-	Id       int   `json:"id"`
+	Id       int64   `json:"id"`
 	Orderno  string  `json:"orderno"`
 	Username string  `json:"username"`
-	Amount   float64 `json:"amount"`
+	Amount   float32 `json:"amount"`
 	Status   string  `json:"status"`
 	Fileurl  string  `json:"fileurl"`
 	Time     string  `json:"time"`
@@ -15,6 +15,6 @@ type DemoOrderList struct {
 }
 
 type Student struct {
-	Id      int    `json:"id"`
+	Id      int64    `json:"id"`
 	Number  string   `json:"number"`
 }
